+++ CHANGELOG.md
@@ -22,7 +22,9 @@
 - `ricos-schema`
   - [#3617](https://github.com/wix/ricos/pull/3617) add lists indentation
 - `toolbars-new/ui-components`
-  - [#3665](https://github.com/wix/ricos/pull/3665) wire icons color 
+  - [#3665](https://github.com/wix/ricos/pull/3665) wire icons color
+- `ricos-viewer`
+  - [#3667](https://github.com/wix/ricos/pull/3667) support print css
 
 </details>
   <hr/>